move timer list into its own component
