+++ boardgame-ui/src/locales/translations.csv
@@ -39,6 +39,7 @@ event.tab.boardgames,Board Games,
 event.tab.games,Games,
 event.tab.rankings,Rankings,
 event.tab.matchmaking,Matchmaking,
+event.tab.timers,Timers,
 event.rankings.games-won,Games won,
 event.rankings.games-won-percentage,Games won (%),
 event.rankings.games-played,Games played,
@@ -183,6 +184,7 @@ timers.add,Create Timer,
 timers.creator.you,you,
 timers.timer,Timer,
 timers.created-by,Created by,
+timers.confirm-timer-deletion,Are you sure you want to delete this game?,
 add-game.title,Add game,
 edit-game.title,Edit game,
 add-edit-game.board-game.label,Board game,
